Relax FocusPoint implementation
Remove FocusPoints Scriptable Object and instead store the list of focus points inside the FocusPointConstraint class.
