--- Runtime/Animation/Rigging/FocusPoints.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f7f69435d6f50384291b42191b3eedaa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
